Referencia antiga/requirements.txt
# ============================================================
# TRIBUNAL GOLDENMASTER - Dependências Python
# ============================================================
# Requer: Python 3.10+
# Instalar: pip install -r requirements.txt
# ============================================================

# ---------------------------------------------------------
# Interface Web
# ---------------------------------------------------------
streamlit>=1.28.0

# ---------------------------------------------------------
# Dados e Validação
# ---------------------------------------------------------
pandas>=2.0.0
pydantic>=2.0.0

# ---------------------------------------------------------
# HTTP e API
# ---------------------------------------------------------
httpx>=0.25.0
tenacity>=8.2.0

# ---------------------------------------------------------
# Leitura de Documentos
# ---------------------------------------------------------
pypdf>=3.17.0
python-docx>=1.0.0
openpyxl>=3.1.0
pdfplumber>=0.10.0

# ---------------------------------------------------------
# PDF Seguro (extração página-a-página)
# ---------------------------------------------------------
pymupdf>=1.23.0
pillow>=10.0.0

# ---------------------------------------------------------
# Web Scraping (DRE - Diário da República)
# ---------------------------------------------------------
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ---------------------------------------------------------
# Export PDF
# ---------------------------------------------------------
reportlab>=4.0.0

# ---------------------------------------------------------
# Ambiente
# ---------------------------------------------------------
python-dotenv>=1.0.0

# ---------------------------------------------------------
# Testes (desenvolvimento)
# ---------------------------------------------------------
pytest>=7.4.0
pytest-cov>=4.1.0

# ---------------------------------------------------------
# OCR (OPCIONAL - requer Tesseract instalado no sistema)
# Descomentar se precisar de OCR para PDFs escaneados
# ---------------------------------------------------------
# pytesseract>=0.3.10
